edit batch doesn't need project ID
